Add files via upload
So im working on one of my lab programs for my class and im having issues debugging my program. If anybody has any insight that would be really helpful. Im on the other side of the country so its almost impossible to be on the same time line as my professor. Current errors are that it will not compile, the variable userCon is not being read, my for statement in case 1 is not being read and my counter string is not being read, as well as userOp.